added section and others requirment

+++ app/Http/Controllers/BundleController.php
@@ -30,18 +30,29 @@ class BundleController extends Controller
         'name' => 'required|max:255',
         ]);
         $data['name'] = $request->name;
+        $data['slug'] = preg_replace('/\s+/', '-', $request->name);
         $data['user_id'] = $user->id;
-        Bundle::create($data);
-        return redirect()->back();
+        $bundle = Bundle::create($data);
+        $cover['name'] = "Cover Page";
+        $cover['bundle_id'] = $bundle->id;
+        $cover['user_id'] = $user->id;
+        $cover['isDefault'] = 1;
+        Section::create($cover);
+        $Index['name'] = "Index";
+        $Index['bundle_id'] = $bundle->id;
+        $Index['user_id'] = $user->id;
+        $Index['isDefault'] = 1;
+        Section::create($Index);
+        return redirect()->route('bundle.show_single', [$bundle->slug,$bundle->id]);
     }
-    public function show($id)
+    public function show($slug,$id)
     {
         $user = Auth::user();
         if ($user->isAdmin()) {
             return view('backend.pages.dashboard');
         }
 
-        $bundle = Bundle::with('section')->where(['user_id'=>$user->id,"id"=>$id])->first();
+        $bundle = Bundle::with('section')->where(['user_id'=>$user->id,"slug"=>$slug,"id"=>$id])->first();
         return view('backend.pages.bundle.show',['bundle'=>$bundle]);
     }
     public function edit($id)
@@ -51,7 +62,7 @@ class BundleController extends Controller
             return view('backend.pages.dashboard');
         }
 
-        $bundle = Bundle::where(['user_id'=>$user->id,"id"=>$id])->first();
+
         return view('backend.pages.bundle.edit',['bundle'=>$bundle]);
     }
     public function update(Request $request,$id)
@@ -61,8 +72,8 @@ class BundleController extends Controller
             return view('backend.pages.dashboard');
         }
         $name = $request->name;
-
-        Bundle::where(['user_id'=>$user->id,"id"=>$id])->update(['name'=>$name]);
+        $slug = preg_replace('/\s+/', '-', $request->name);
+        Bundle::where(['user_id'=>$user->id,"id"=>$id])->update(['name'=>$name,'slug'=>$slug]);
         return redirect()->route('bundle.index');
     }
     public function destroy($id)
@@ -75,8 +86,10 @@ class BundleController extends Controller
             foreach($sec as $s)
             {
                 $file = File::where('section_id',$s->id)->first();
-                unlink(public_path($file->filename));
-                File::where('id',$file->id)->delete();
+                if(!is_null($file)){
+                    unlink(public_path("pdf/".$file->filename));
+                    File::where('id',$file->id)->delete();
+                }
                 Section::where('id',$s->id)->delete();
             }
             $bundle->delete();
@@ -91,7 +104,7 @@ class BundleController extends Controller
         if($bundle->count() > 0)
         {
             $file = $bundle->first();
-                unlink(public_path($file->filename));
+                unlink(public_path("generated_pdf/".$file->filename));
                 generatedTable::where('id',$file->id)->delete();
             return redirect()->back();
         }else{
